CMakeLists: Don't set VCTR_USE_IPP compile definition if according CMake option is not enabled
The VCTR_USE_IPP CMake option enables automatic linkage of IPP via the VCTR CMakeLists. There are situations, where the user might want to use IPP but wants to link to it in a different way and thus opts to not set that CMake option. Manually setting the compile definition should still be possible in that case.